backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.12.0
supabase==2.3.4
python-dotenv==1.0.0
httpx==0.26.0
Pillow==10.2.0
aiofiles==23.2.1
